system76: Remove kernel param ec_sys.write_support=1.
I don't think we need this in the generic configuration.  There is a call to
`modprobe ec_sys.write_support=1` in system76-driver, only for the airplane mode
daemon (which only applies to certain models, not including darp6), and the
driver executes the modprobe call anyhow.

Also the kernel's documentation for the parameter says:

    write_support:Dangerous, reboot and removal of battery may be needed. (bool)

So let's only enable this where it's required.