lenovo/thinkpad/x260: i915.enable_psr=0 parameter
The video driver causes short freezes from time to time, especially if
an external monitor is attached. Adding the i915.enable_psr=0 kernel
parameter mitigates those freezes.

This is a known problem for the X260, as mentioned in the Arch Wiki. It
occurred mainly after switching to kernel version five.