Override edid of built-in display
Fix the problem that the edid of the built-in display only provided 60hz mode when use Hybrid Graphics

It appears to be a Lenovo firmware issue and i try to fix it. This issue still reproducible on BIOS version GKCN58WW(22/12/16)
Lenovo Legion 5 Pro is a laptop which support a technology called "DDG" that can allow you switch between discrete graphics mode and hybrid mode (optiums).
In discrete graphics mode, the firmware provides the correct edid and anything works fine, the built-in display work well at 165hz.
But if switch to hybrid mode, the firmware provides a different edid, causing the built-in display only can work at 60hz.
So I extracted the edid file of discrete graphics mode and override the edid that built-in display provide to solve this problem.

Co-Authored-By: Jörg Thalheim <Mic92@users.noreply.github.com>