common-gpu-nvidia: expose architecture modules in `nixosModules`
The architecture-specific modules for `common-gpu-nvidia` are used in
the various hardware modules for specific hardware depending on what is
needed.

This works well, but it would be nice to have these exposed through the
flake `nixosModules` so that users using "generic" systems can still
access these modules in the same manner as all the others. Such as in my
case, I have a desktop system running an Nvidia Ampere GPU, but have to
access the `common/gpu/nvidia/ampere` module through its path since it
is not exposed in the flake itself.

I did not add these to the `README` since none of the other
`common-gpu-nvidia` modules are presented there. I opted to follow the
trend. I can add them all if suggested.