lenovo/thinkpad/x13s: source kernel through `config` instead of `pkgs`
The previous implementation was sourcing the kernel through `pkgs.linux`,
which is only representative of the final system if `boot.kernelPackages`
is left as the default value of `pkgs.linuxPackages`.

You can of course change this to other package sets, such as
`pkgs.linuxPackages_latest`. Instead, we now reference the kernel
through `config.boot.kernelPackages.kernel`.